+++ .changeset/grumpy-wombats-talk.md
@@ -1,0 +1,5 @@
+---
+"@sumup-oss/circuit-ui": minor
+---
+
+Added a new hook `useScrollLock` to disable page scroll on demand.
