+++ packages/circuit-ui/components/Modal/ModalService.spec.tsx
@@ -13,11 +13,9 @@
  * limitations under the License.
  */
 
-import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { describe, expect, it } from 'vitest';
 
-import { renderHook, act } from '../../util/test-utils.js';
-
-import { getKeyboardFocusableElements, useScrollLock } from './ModalService.js';
+import { getKeyboardFocusableElements } from './ModalService.js';
 
 describe('DialogService', () => {
   describe('getKeyboardFocusableElements', () => {
@@ -85,71 +83,6 @@ describe('DialogService', () => {
           container,
         ]),
       );
-    });
-  });
-  describe('useScrollLock', () => {
-    Object.defineProperty(window, 'scrollTo', {
-      value: vi.fn(),
-      writable: true,
-    });
-
-    Object.defineProperty(window, 'scrollY', { value: 1, writable: true });
-
-    beforeEach(() => {
-      document.body.style.position = '';
-      document.body.style.top = '';
-      document.documentElement.style.setProperty('--scroll-y', '');
-    });
-
-    it('locks the scroll when `isLocked` is true', () => {
-      document.documentElement.style.setProperty('--scroll-y', '100px');
-
-      const { rerender } = renderHook(
-        ({ isLocked }) => useScrollLock(isLocked),
-        {
-          initialProps: { isLocked: false },
-        },
-      );
-
-      rerender({ isLocked: true });
-
-      expect(document.body.style.position).toBe('fixed');
-      expect(document.body.style.top).toBe('-100px');
-    });
-
-    it('unlocks the scroll when `isLocked` is false', () => {
-      document.body.style.top = '-100px';
-
-      const { rerender } = renderHook(
-        ({ isLocked }) => useScrollLock(isLocked),
-        {
-          initialProps: { isLocked: true },
-        },
-      );
-
-      rerender({ isLocked: false });
-
-      expect(document.body.style.position).toBe('');
-      expect(document.body.style.top).toBe('');
-      expect(window.scrollTo).toHaveBeenCalledWith(0, 100);
-    });
-
-    it('updates `--scroll-y` on scroll', () => {
-      global.requestAnimationFrame = vi
-        .fn()
-        .mockImplementation((callback: () => void) => callback());
-
-      renderHook(() => useScrollLock(false));
-
-      act(() => {
-        window.scrollY = 200;
-        const scrollEvent = new Event('scroll');
-        window.dispatchEvent(scrollEvent);
-      });
-
-      expect(
-        document.documentElement.style.getPropertyValue('--scroll-y'),
-      ).toBe('200px');
     });
   });
 });
