+++ packages/circuit-ui/index.ts
@@ -222,3 +222,4 @@ export { useFocusList } from './hooks/useFocusList/index.js';
 export { useCollapsible } from './hooks/useCollapsible/index.js';
 export { useSwipe } from './hooks/useSwipe/index.js';
 export { useMedia } from './hooks/useMedia/index.js';
+export { useScrollLock } from './hooks/useScrollLock/useScrollLock.js';
